shared code and guid neutral
